refactor, cleanup, fix TextVal extra quotes bug
