+++ src/main/kotlin/senior/joinu/candid/idl/IDLParser.kt
@@ -1,4 +1,4 @@
-package senior.joinu.candid
+package senior.joinu.candid.idl
 
 import com.github.h0tk3y.betterParse.combinators.*
 import com.github.h0tk3y.betterParse.grammar.Grammar
@@ -6,6 +6,7 @@ import com.github.h0tk3y.betterParse.grammar.parser
 import com.github.h0tk3y.betterParse.lexer.literalToken
 import com.github.h0tk3y.betterParse.lexer.regexToken
 import com.github.h0tk3y.betterParse.parser.Parser
+import senior.joinu.candid.utils.idlHash
 import java.util.regex.Pattern
 
 
@@ -297,6 +298,7 @@ sealed class IDLToken {
 
     data class TextVal(override val value: String) : IDLToken(), IDLName {
         companion object { val pattern: String = Pattern.compile("\"[\\w._\\-\\\\/:]+\"", Pattern.UNICODE_CHARACTER_CLASS).pattern() }
-        override fun toString() = "\"$value\""
+
+        override fun toString() = value
     }
 }
