+++ src/main/kotlin/senior/joinu/candid/idl/IDLType.kt
@@ -1,9 +1,10 @@
-package senior.joinu.candid
+package senior.joinu.candid.idl
 
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.asTypeName
 import senior.joinu.candid.serialize.getTypeSerForType
 import senior.joinu.candid.transpile.prettyString
+import senior.joinu.candid.utils.Leb128
 import java.nio.ByteBuffer
 
 class TypeTable(
@@ -378,7 +379,11 @@ sealed class IDLDef {
 }
 
 data class IDLActor(val name: String?, val type: IDLActorType) {
-    override fun toString() = "${IDLType.Reference.Service.text} ${if (name == null) "" else "$name"}:${type.toString().replace(IDLType.Reference.Service.text, "")};"
+    override fun toString() = "${IDLType.Reference.Service.text} ${if (name == null) "" else "$name"}:${
+        type.toString().replace(
+            IDLType.Reference.Service.text, ""
+        )
+    };"
 }
 
 data class IDLProgram(val imports: List<IDLDef.Import>, val types: List<IDLDef.Type>, val actor: IDLActor?) {
