+++ src/main/kotlin/senior/joinu/candid/serialize/TypeSers.kt
@@ -1,8 +1,8 @@
 package senior.joinu.candid.serialize
 
 import com.squareup.kotlinpoet.CodeBlock
-import senior.joinu.candid.*
-import senior.joinu.candid.Leb128
+import senior.joinu.candid.idl.*
+import senior.joinu.candid.utils.Leb128
 import java.nio.ByteBuffer
 import java.nio.charset.StandardCharsets
