+++ src/main/kotlin/senior/joinu/candid/transpile/Complex.kt
@@ -2,9 +2,15 @@ package senior.joinu.candid.transpile
 
 import com.squareup.kotlinpoet.*
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
-import senior.joinu.candid.*
+import senior.joinu.candid.idl.IDLFuncAnn
+import senior.joinu.candid.idl.IDLType
+import senior.joinu.candid.idl.MAGIC_PREFIX
+import senior.joinu.candid.idl.TypeTable
 import senior.joinu.candid.serialize.*
-import senior.joinu.candid.Leb128
+import senior.joinu.candid.utils.EdDSAKeyPair
+import senior.joinu.candid.utils.Leb128
+import senior.joinu.candid.utils.escapeIfNecessary
+import senior.joinu.candid.utils.poetize
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
 import java.util.*
