+++ src/test/kotlin/senior/joinu/candid/parser/ExampleService.kt
@@ -1,16 +1,11 @@
-import kotlinx.coroutines.runBlocking
-import java.nio.ByteBuffer
-import java.nio.ByteOrder
-import java.util.Base64
-import kotlin.ByteArray
-import kotlin.String
-import senior.joinu.candid.EdDSAKeyPair
-import senior.joinu.candid.SimpleIDLFunc
-import senior.joinu.candid.SimpleIDLPrincipal
-import senior.joinu.candid.SimpleIDLService
 import senior.joinu.candid.serialize.FuncValueSer
 import senior.joinu.candid.serialize.ServiceValueSer
 import senior.joinu.candid.serialize.TypeDeser
+import senior.joinu.candid.transpile.SimpleIDLFunc
+import senior.joinu.candid.transpile.SimpleIDLService
+import java.nio.ByteBuffer
+import java.nio.ByteOrder
+import java.util.*
 
 typealias MainActorValueSer = ServiceValueSer
 
@@ -37,32 +32,10 @@ class AnonFunc0(
         receiveBuf.put(receiveBytes)
         receiveBuf.rewind()
         val deserContext = TypeDeser.deserUntilM(receiveBuf)
-        return senior.joinu.candid.serialize.TextValueSer.deser(receiveBuf) as kotlin.String
+        return senior.joinu.candid.serialize.TextValueSer.deser(receiveBuf)
     }
 
     companion object {
         val staticPayload: ByteArray = Base64.getDecoder().decode("RElETAABcQ==")
-    }
-}
-
-class MainActor(
-    host: String,
-    canisterId: SimpleIDLPrincipal?,
-    keyPair: EdDSAKeyPair?,
-    apiVersion: String = "v1"
-) : SimpleIDLService(host, canisterId, keyPair, apiVersion) {
-    val greet: AnonFunc0 = AnonFunc0("greet", this)
-}
-
-fun main() {
-    runBlocking {
-        val id = SimpleIDLPrincipal.fromText("7kncf-oidaa-aaaaa-aaaaa-aaaaa-aaaaa-aaaaa-q")
-        val keyPair = EdDSAKeyPair.generateInsecure()
-        val host = "http://localhost:8000"
-
-        val actor = MainActor(host, id, keyPair)
-
-        val response = actor.greet("joinu")
-        println(response)
     }
 }
